requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Redis
redis[asyncio]==5.0.4

# RabbitMQ
aio-pika==9.4.1

# HTTP client (for inter-service calls in future)
httpx==0.27.0
